Fix ad request muted state
The ad request ignores if the player is in muted state, accounting only if the player can autostart with audio.
This will cause policy violations, because the ad request will not reflect the real audibility of the ad placement.

This fix addresses the problem, by taking into account the actual muted state of the player.